### 2023-09-11 #5834 @editorialbot
[PRE REVIEW]: genderstat: An R Package for Comprehensive Gender Inequality Analysis
**Submitting author:** @mashrur-ayon (<a href="http://orcid.org/0000-0002-3659-2891">S M Mashrur Arafin Ayon</a>)
**Repository:** https://github.com/mashrur-ayon/genderstat
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.3
