### 2023-09-11 #5834 @mashrur-ayon
Dear @spholmes,

Thank you for taking the time to review the genderstat package and for your constructive feedback. I truly respect and value your input.

I am aware of the retirement of the `reshape2` package and appreciate you highlighting this dependency concern. As you correctly pointed out, `tidyr` is now the recommended replacement for `reshape2`. Given that version 0.1.3 of `genderstat` has already been approved on **CRAN**, I will be sure to take your suggestion into account and transition to using `tidyr` in the upcoming version 0.1.4.

I understand the importance of ensuring stability and consistency for the users. Therefore, I will take utmost care to ensure that this shift from `reshape2` to `tidyr` does not adversely affect the user experience or the package's functionality.

Once again, thank you very much for pointing this out. Your feedback is integral in improving the quality and reliability of the genderstat package.

_Warm regards,_
Mashrur.
