### 2023-09-11 6df2892
Update README.md
Included JOSS submission badge.
