+++ paper/reference.bib
@@ -36,14 +36,13 @@
   organization = {R-Packages}
 }
 
-@article{wickham_2007_reshaping,
-  author = {Wickham, Hadley},
-  title = {Reshaping Data with thereshapePackage},
-  doi = {10.18637/jss.v021.i12},
-  volume = {21},
-  year = {2007},
-  journal = {Journal of Statistical Software}
+@Manual{wickham_2007_reshaping,
+  title = {tidyr: Tidy Messy Data},
+  author = {Hadley Wickham and Davis Vaughan and Maximilian Girlich},
+  year = {2023},
+  note = {https://tidyr.tidyverse.org, https://github.com/tidyverse/tidyr},
 }
+
 @Book{wickham_2016,
     author = {Hadley Wickham},
     title = {ggplot2: Elegant Graphics for Data Analysis},
