### 2023-09-11 31188bb
Update paper.md
updated code to install from github. Removed the dependency of reshape2 package.
