### 2023-09-11 1698c1f
Update plot_gem.R
