### 2023-09-11 376707f
Update plot_gdi.R
