### 2023-09-11 1d1bdf4
updating package to 0.1.4
removed the reshape2 dependency to tidyr
