+++ genderstat/NAMESPACE
@@ -18,7 +18,7 @@ export(plot_gpg_q3)
 export(plot_gpg_q4)
 export(plot_hdi)
 import(dplyr)
-import(reshape2)
+import(tidyr)
 importFrom(ggplot2,aes)
 importFrom(ggplot2,coord_flip)
 importFrom(ggplot2,element_text)
