+++ paper/reference.bib
@@ -40,7 +40,8 @@
   title = {tidyr: Tidy Messy Data},
   author = {Hadley Wickham and Davis Vaughan and Maximilian Girlich},
   year = {2023},
-  note = {https://tidyr.tidyverse.org, https://github.com/tidyverse/tidyr},
+  url = {https://tidyr.tidyverse.org},
+  note = {https://github.com/tidyverse/tidyr},
 }
 
 @Book{wickham_2016,
