### 2023-09-11 d722b22
Update paper.md
Tested the installing package from github
