### 2023-09-11 7168f6b
Update paper.md
Fixed issue with github url
