+++ paper/paper.md
@@ -111,7 +111,8 @@ install.packages("dplyr")
 install.packages("readr")
 install.packages("genderstat")
 
-# If you want to obtain the latest version of genderstat package from github, use this code instead
+# If you want to obtain the latest version of genderstat package from GitHub.
+# Use this code instead
 install.packages("devtools")
 devtools::install_github("mashrur-ayon/genderstat", subdir = "genderstat")
 
