### 2023-09-11 #5834 @mashrur-ayon
Dear @spholmes and @oliviaguest,

I hope this message finds you well. I wanted to express my gratitude to **Susan** for pointing out the retirement of `reshape2`. Taking your feedback into account, I've promptly made the necessary edits to my R code. I've updated the package on GitHub, revised the paper, and adjusted the references accordingly.

As a graduate in gender studies and an ardent supporter of open-source initiatives, I'm deeply committed to the continuous improvement and maintenance of the `genderstat` package. My dedication to the project motivated me to address the `reshape2` issue and transition to `tidyr`, ensuring the package's longevity and compatibility. While genderstat was functional with `reshape2`, I recognize the importance of staying updated with current best practices. I genuinely appreciate your guidance in this matter, **Susan**.

Thank you both for your time and invaluable feedback.

_Warm regards,_
Mashrur.
