### 2023-10-09 #5834 @spholmes
👋  Maria @thegargiulian
Would you be interested in reviewing this for JOSS.
Many thanks
Susan
