### 2023-10-12 #5932 @joaojcorreia
The paper should have a clearer summary, focusing on the library's purpose and its original data sources. The statement of need needs to better explain the package's benefits and be more transparent about data sources. The article lacks context about the field and how GenderStat fits in. The writing style should be simpler, and function descriptions clearer. The documentation should provide more detail on example usage, parameter explanations, and function functionality. It should also clarify data origin and include a link to CRAN documentation.

Bellow you will find my detailed review notes.
________________________________________________________________

**Notes on the article:**

Summary:
- Summary is unclear on the purpose of the library.
- Summary is the only place that references the example dataset. This dataset is compiled by the author from other sources, so the emphasis should be put on the original sources of the data, and how they were processed and aggregated.
- Too much emphasis is put on the fact that this is open source software and its benefits. This is a JOSS paper, so there is no need to "preach to the choir."
- More emphasis should be put on the metrics and visualizations that the package generates, and their usefulness in research and pedagogical settings.

Statement of need:
- The language is unclear.
- Can the package be used to display data from other sources? (The documentation says yes, but this is not clear in the statement of need.)
- Why is the package needed? Does it make researchers' tasks simpler? Is it a one-stop shop for researchers to gather gender inequality statistics and display them? If so, more transparency is needed on the data sourcing and compiling process.

Assessment of the state of the field:
- The author does not make an effort to assess the state of the field and how the tool fits into it.
- I would suggest a brief paragraph or two to highlight the state of research in this field and how genderstat fits into it.

Quality of the writing:
- The quality of the writing needs to be improved.
- The author occasionally uses overcomplex and convoluted expressions. I would recommend a simpler and more direct writing style.
- The descriptions of the functions themselves are insufficient.

References:
- The references seem complete and appropriately cited. However, the article would be improved if more context were provided about the indexes being presented in the Core theoretical concepts section. This context could include:
  - Who originally developed the concept?
  - What are some relevant examples of where the index has been used?
- Providing this additional context would help readers to better understand the significance of the indexes and how they are used in the field.

General comments:
- Figure 1 helps understand how the library is structured, and what it is for, yet I am sure the author could figure out a better ways of representing it.

**Notes on the documentation:**

General comments:
- References the origin of the data, yet it should either linked or give further details on its origin. The data is based on official statistics from the UNDP and World Bank. Yet which part is which? The data refers to which year? How was the data aggregated?
- No link to the CRAN documentation on the README in the Git repository
- ‘gender_development_index(data)’ does not work

Statement of need:
- Same as above.

Installation instructions:
- They are clear and straight forward.

Example usage:
- They aren’t clear and detailed.
- The author should include details on how the data-structure for import data.
- No detail and examples are given on parameters for the listed functions. E.g. does ‘plot_gdi(real_data_GDI)’ always plot the bottom 15 countries?

Functionality documentation:
- More detail needs to be provided about the functions.

Automates tests:
- Seem to be missing.

Community guidelines:
- Please expand. It only has a brief statement about contributions.
