### 2023-11-11 mashrur-ayon/genderstat#5 @thegargiulian
Installation instructions
I'd recommend simplifying the installation instructions to just `install.packages("genderstat")`.

You're already managing the `ggplot2` and `tidyverse` dependencies automatically in the `DESCRIPTION` file, so there's no need to include those dependencies in the installation instructions. The same goes for the dependency description you have in the README.

_This issue is with reference to the JOSS review here https://github.com/openjournals/joss-reviews/issues/5932_
