### 2023-11-11 mashrur-ayon/genderstat#6 @thegargiulian
Add testing and automate test running
You can use `testthat` to develop tests to ensure that the calculations for the various indices you include are correct and then automate those tests with continuous integration (more [here](https://usethis.r-lib.org/reference/github_actions.html)).

*This issue is with reference to the JOSS review here https://github.com/openjournals/joss-reviews/issues/5932*
