### 2024-02-04 #5932 @spholmes
> Hi @mashrur-ayon, I hope you're well!
>
> I've opened a couple of issues related to specific parts of the package. I agree largely with @joaojcorreia's comments, in particular, the paper and the package would benefit from reframing to highlight the project's main contribution—the data itself.
>
> Notes about the paper:
>
> * The paper and documentation should focus more on the provenance of the data: where did the data come from?, how and why was it created?, how was it processed by you for the purposes of this package? I would also recommend more carefully motivating _why_ we need this package because that is not entirely clear from the statement of need in its current form.
> * The paper needs a more thorough survey of the field. While `genderstat` might be the first R package dedicated to making metrics related to the study of gender (in)equality, the paper would benefit from the discussion other attempts to elucidate these issues, even if they haven't resulted in the publication of an open source R package.
> * It would be helpful to provide more information about the core theoretical concepts you mention, especially how they have been applied or used, especially in social science research in the past. I'd also recommend a short discussion of why you chose these particular metrics as other metrics of gender equality do exist.
> * Generally, the language used in the paper is often overly complex and should be simplified. This should improve overall readability and reduce the word count.
>
> Notes about the package:
>
> * I would consider restructuring or eliminating the graphing functions from the package. In their current form, the graphing functions do not provide users any flexibility to customize the graphics in the way that most suits their needs (e.g., `plot_gdi()` always plots the bottom 15 countries according to GDI). Instead of offering users these pre-built graphing functions, I would instead suggest providing some simple worked examples of real world social science research questions that users would be able to examine graphically using the data provided in the `genderstat` package. These types of examples would also improve the overall quality of the documentation.
> * Some of the datasets/metrics have missing values in them. I'd recommend commenting on this in the data or function documentation.
> * The data does not provide any time information, e.g., what year are the observations from? Without this information there is little context for the data. Additionally ,it appears that data is only available for a single year. I think the scholarly contribution of this package would be augmented by including a longer time series of information from the different sources you draw from. This would also provide crucial functionality for users who may be interested in how these metrics change over time.

Hi @mashrur-ayon , I wondered whether you could respond to the revisions requested by the reviewers?
If you are not able to respond, how would like to proceed, do you want to withdraw at this time?
Thanks,
Susan
