### 2024-02-05 #5932 @spholmes
> Agreed @arfon , I was waiting for the author to step in and make the changes requested by the reviewers, I'll ping them now to see where they stand on either making the changes or withdrawing the paper.

Hi @arfon, I did get a reply from the author and he has promised to make progress on the revisions.
Thanks
Susan
