+++ genderstat/DESCRIPTION
@@ -1,9 +1,9 @@
 Package: genderstat
 Type: Package
 Title: Quantitative Analysis Tools for Gender Studies
-Version: 0.1.4
+Version: 0.1.5
 Authors@R: person("S M Mashrur", "Arafin Ayon", email = "mashrur399@gmail.com", role = c("aut", "cre"), comment = c(ORCID = "0000-0002-3659-2891"))
-Description: Provides tools for quantitative analysis in gender studies, including functions to calculate various gender inequality metrics such as the Gender Pay Gap, Gender Inequality Index (GII), Gender Development Index (GDI), and Gender Empowerment Measure (GEM). Also includes extracted real datasets for practice and learning purposes, which were obtained from the UNDP Human Development Reports Data Center <https://hdr.undp.org/data-center/documentation-and-downloads> and the World Bank Gender Data Portal <https://genderdata.worldbank.org/indicators/>. References: Miller, Kevin; Vagins, Deborah J. (2021) <https://eric.ed.gov/?id=ED596219>. Jacques Charmes & Saskia Wieringa (2003) <doi:10.1080/1464988032000125773>. Gaëlle Ferrant (2010) <https://shs.hal.science/halshs-00462463/>.
+Description: Provides tools for quantitative analysis in gender studies, including functions to calculate various gender inequality metrics such as the Gender Pay Gap, Gender Inequality Index (GII), Gender Development Index (GDI), and Gender Empowerment Measure (GEM). Also includes extracted secondary example datasets for practice and learning purposes, which were obtained from the UNDP Human Development Reports Data Center and the World Bank Gender Data Portal by the author the dataset is available on <doi:10.34740/kaggle/dsv/6359326>. References: Miller, Kevin; Vagins, Deborah J. (2021) <https://eric.ed.gov/?id=ED596219>. Jacques Charmes & Saskia Wieringa (2003) <doi:10.1080/1464988032000125773>. Gaëlle Ferrant (2010) <https://shs.hal.science/halshs-00462463/>.
 License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
