### 2024-07-06 224ebaa
package restructered
removed all the plot functions for better structure and scope of the package.
