+++ paper/reference.bib
@@ -85,3 +85,40 @@ url = {https://doi.org/10.7910/DVN/GZLYLI}
   year = {2021},
   organization = {cran.r-project.org}
 }
+@article{blau_2017_gender,
+  author = {Blau, Francine D. and Kahn, Lawrence M.},
+  title = {The Gender Wage Gap: Extent, Trends, and Explanations},
+  journal = {Journal of Economic Literature},
+  volume = {55},
+  number = {3},
+  pages = {789-865},
+  year = {2017},
+  doi = {10.1257/jel.20160995}
+}
+
+@book{grown_2005_taking,
+  author = {Grown, C. and Gupta, G. R. and Pande, R.},
+  title = {Taking Action: Achieving Gender Equality and Empowering Women},
+  publisher = {UN Millennium Project Task Force on Education and Gender Equality},
+  year = {2005},
+  doi = {10.4324/9781849773560}
+}
+
+@article{buchmann_2008_gender,
+  author = {Buchmann, C. and DiPrete, T. A. and McDaniel, A.},
+  title = {Gender Inequalities in Education},
+  journal = {Annual Review of Sociology},
+  volume = {34},
+  pages = {319-337},
+  year = {2008},
+  doi = {10.1146/annurev.soc.34.040507.134719}
+}
+
+@book{paxton_2014_women,
+  author = {Paxton, P. and Hughes, M. M.},
+  title = {Women, politics, and power: A global perspective},
+  publisher = {SAGE Publications, Inc.},
+  year = {2014},
+  doi = {10.4135/9781452275482}
+}
+
