+++ paper/reference.bib
@@ -121,4 +121,11 @@ url = {https://doi.org/10.7910/DVN/GZLYLI}
   year = {2014},
   doi = {10.4135/9781452275482}
 }
-
+@misc{s_m_mashrur_arafin_ayon_2023,
+	title={Gender Metrics by Country: Socio-Economic &amp; Health},
+	url={https://www.kaggle.com/dsv/6359326},
+	DOI={10.34740/KAGGLE/DSV/6359326},
+	publisher={Kaggle},
+	author={S M Mashrur Arafin Ayon},
+	year={2023}
+}
