+++ paper/paper.md
@@ -22,15 +22,32 @@ affiliations:
 
 # Summary
 
-The `genderstat` package is an exhaustive tool developed for the `R` [@R] programming environment, explicitly devised to expedite quantitative evaluations in the field of gender studies. It encompasses an array of functions that facilitate the computation of pivotal gender disparity indicators, including the _Gender Pay Gap_ **(GPG)**, _Gender Inequality Index_ **(GII)**, _Gender Development Index_ **(GDI)**, and _Gender Empowerment Measure_ **(GEM)**. Moreover, it incorporates authentic datasets[@DVN/GZLYLI_2023], derived from the UNDP Human Development Reports Data Center and the World Bank Gender Data Portal, which can be leveraged for pedagogical and practical applications.
+The `genderstat` package is a comprehensive tool developed for the R programming environment, designed to facilitate quantitative evaluations in gender studies. It includes functions to calculate essential gender disparity indicators such as the  _Gender Pay Gap_ **(GPG)**, _Gender Inequality Index_ **(GII)**, _Gender Development Index_ **(GDI)**, and _Gender Empowerment Measure_ **(GEM)**. These indicators are crucial as they provide a clear picture of gender disparities across various dimensions, including economic participation, health, education, and empowerment. Specifically, **GII**, **GDI**, and **GEM** encapsulate complex aspects of gender inequality that are not addressed by single metrics. 
 
-This package is fundamentally engineered to support scholars in the social sciences engaged in gender-centric research. Being a freely accessible open-source resource from `R` [@R], it does not incur any financial burdens. Moreover, by being open-source, it supports the ethos of collaborative knowledge sharing and innovation that underpin open-source initiatives and their core philosophy. This provision enables researchers to focus more intently on their profound theoretical explorations rather than dedicating an exorbitant amount of time to craft functions for data visualization or resorting to expensive proprietary alternatives. The `genderstat` package is endowed with the requisite functionalities to scrutinize and visualize gender metrics, thereby serving as a specialized asset for gender analysis.
+The package also provides example datasets sourced from the UNDP Human Development Reports Data Center and the World Bank Gender Data Portal. Only the relevant portions of these datasets have been collected and maintained on the author's Kaggle account as a consolidated dataset[@s_m_mashrur_arafin_ayon_2023], which is updated annually. These example datasets are integral to the package, serving both educational and practical research purposes. They are utilized to build vignettes and showcase examples of how to use the package effectively.
 
-Additionally, the genderstat package relies on other `R` [@R] packages, namely `ggplot2` [@wickham_2016], `dplyr` [@wickham_2020_dplyr], and `tidyr` [@wickham_2007_reshaping], to provide its full range of functionalities. These dependencies are crucial for the data manipulation, transformation, and visualization tasks that `genderstat` performs.
+`genderstat` simplifies these computations, making them accessible and easy to use. This package streamlines the process, allowing social scientists, particularly those conducting research on gender studies, to perform complex analyses with minimal coding effort, thereby saving time and reducing the potential for errors. It is intended to simplify the work of gender studies researchers by providing a one-stop solution for measuring various gender indexes. It allows R programmers, especially those who want to engage in gender studies research, to easily analyze and visualize gender metrics without needing to create these functions from scratch.
+
+By streamlining the process of gender disparity analysis, `genderstat` enables researchers to focus more on their theoretical and practical explorations rather than the technical aspects of data analysis.
 
 # Statement of Need
 
[…]
