### 2024-07-07 5eb63e4
updated paper
updated reference, figure and output table
