+++ paper/paper.md
@@ -48,23 +48,26 @@ Recent studies highlight the persistent gender pay gap and the underrepresentati
 
 
 
+## How `genderstat` Fits into the Existing Research Landscape
+
+The `genderstat` package addresses the growing need for quantitative tools in gender studies by providing researchers with a comprehensive suite of functions to analyze gender disparity indicators. By simplifying the calculation of complex metrics such as the Gender Pay Gap (GPG), Gender Inequality Index (GII), Gender Development Index (GDI), and Gender Empowerment Measure (GEM), `genderstat` enhances the ability of researchers to empirically validate their theoretical findings.
+
+`genderstat` facilitates rigorous empirical analysis, which is crucial for validating theoretical claims in gender studies. For example, it allows researchers to quantify the gender pay gap across different sectors and regions, thereby providing researchers an opportunity to bring out concrete evidence to support policy recommendations. The package also supports the analysis of gender-specific health outcomes, which is essential for developing gender-sensitive health policies. By using `genderstat`, researchers can easily calculate and visualize disparities in health metrics, aiding in the advocacy for more equitable health interventions. Furthermore, `genderstat` aids in the analysis of gender disparities in education and economic outcomes. The package also enables the analysis of gender disparities in political representation, providing insights into the effectiveness of gender quotas and other measures aimed at increasing women's participation in politics. This empirical evidence is crucial for advocating for more inclusive political systems.
+
+By integrating these functionalities, `genderstat` fills a significant gap in the tools available to gender studies researchers, offering a robust, open-source solution that streamlines the process of quantitative gender analysis. It empowers researchers to produce high-quality, empirical evidence that can drive meaningful social change and policy development.
 
 
 # Core Theoretical Concepts
 
 ### Gender Pay Gap: 
 
-The Gender Pay Gap [@miller_2018_the] is a significant metric that encapsulates the differential in the mean earnings of men and women, typically expressed as a percentage. It serves as a lens through which the variance in the remunerations of men and women in the workforce is discerned. A larger percentage denotes a more pronounced incongruity in the earnings of the two genders, while a 0% pay gap symbolizes absolute parity in compensation. As a pivotal barometer of gender inequality, the gender pay gap is shaped by a multitude of determinants, encompassing not merely discrimination, but also disparities in the nature of occupations, the quantum of hours worked, and the opportunities for career advancement[@miller_2018_the].
-
-The Gender Pay Gap is computed employing the subsequent formula:
+The Gender Pay Gap [@miller_2018_the] is a crucial metric that highlights the difference in average earnings between men and women, usually expressed as a percentage. It provides a clear view of the disparities in compensation within the workforce. A higher percentage indicates a greater disparity, while a 0% pay gap signifies equal pay. This gap reflects not only direct discrimination but also differences in the types of jobs held, the number of hours worked, and career advancement opportunities available to men and women. The formula to calculate the Gender Pay Gap is:
 
 $$\text{GPG} = \left( \frac{\text{Male Income} - \text{Female Income}}{\text{Male Income}} \right) \times 100 \%$$
 
 ### Gender Inequality Index:
 
[…]
