+++ paper/reference.bib
@@ -26,32 +26,6 @@
   year = {2010},
   journal = {shs.hal.science}
 }
-
-@misc{wickham_2020_dplyr,
-  author = {Wickham, Hadley and François, Romain and Henry, Lionel and Müller, Kirill and RStudio},
-  month = {08},
-  title = {dplyr: A Grammar of Data Manipulation},
-  url = {https://cran.r-project.org/web/packages/dplyr/index.html},
-  year = {2020},
-  organization = {R-Packages}
-}
-
-@Manual{wickham_2007_reshaping,
-  title = {tidyr: Tidy Messy Data},
-  author = {Hadley Wickham and Davis Vaughan and Maximilian Girlich},
-  year = {2023},
-  url = {https://tidyr.tidyverse.org},
-  note = {https://github.com/tidyverse/tidyr},
-}
-
-@Book{wickham_2016,
-    author = {Hadley Wickham},
-    title = {ggplot2: Elegant Graphics for Data Analysis},
-    publisher = {Springer-Verlag New York},
-    year = {2016},
-    isbn = {978-3-319-24277-4},
-    url = {https://ggplot2.tidyverse.org},
-  }
 @Manual{R,
     title = {R: A Language and Environment for Statistical Computing},
     author = {{R Core Team}},
@@ -60,16 +34,6 @@
     year = {2023},
     url = {https://www.R-project.org/},
   }
-@data{DVN/GZLYLI_2023,
-author = {Arafin Ayon, S M Mashrur},
-publisher = {Harvard Dataverse},
-title = {{Example Dataset for R Package 'genderstat': Gender Disparities Across the Globe - Socio-Economic and Health Metrics}},
-UNF = {UNF:6:LYqx3fk6+zXcTGnhXYY7RA==},
-year = {2023},
-version = {V1},
-doi = {10.7910/DVN/GZLYLI},
-url = {https://doi.org/10.7910/DVN/GZLYLI}
-}
 @misc{conceio_2022_human,
   author = {Conceição, Pedro  and Calderón, Cecilia  and Esbry, Fernanda Pavez  and Ghorai, Moumita  and Hsu, Yu-Chieh  and Ismail, Ghida  and Lengfelder, Christina  and Lutz, Brian  and Mirza, Tasneem  and Mohammed, Rehana  and Pasanen, Josefin  and Shrestha, Som Kumar  and Tapia, Heriberto  and Vázquez, Carolina Rivera  and Yokoi, Yuko  and Zhang,  Yanchun },
   annote = {The equations from the technical notes section of the report were taken in the core theoretical section. Available at: https://hdr.undp.org/sites/default/files/2021-22_HDR/hdr2021-22_technical_notes.pdf},
