### 2024-07-07 e1fa78d
Update README.md
updated readme file as per the version 0.1.5
