+++ README.md
@@ -2,105 +2,121 @@
 
 
 
-`genderstat` package is a comprehensive resource for anyone looking to delve into quantitative analysis in gender studies. It encompasses a variety of functions designed to compute key gender inequality metrics such as the Gender Pay Gap, Gender Inequality Index (GII), Gender Development Index (GDI), and Gender Empowerment Measure (GEM). In addition to these calculation tools, the package also includes real datasets extracted from the UNDP Human Development Reports Data Center and the World Bank Gender Data Portal. These datasets, available for practice and learning, were sourced from reliable databases, thereby ensuring the accuracy and relevance of the analyses performed using this toolkit. 
+The `genderstat` package is a comprehensive tool developed for the R programming environment, designed to facilitate quantitative evaluations in gender studies. It includes functions to calculate essential gender disparity indicators such as the Gender Pay Gap (GPG), Gender Inequality Index (GII), Gender Development Index (GDI), and Gender Empowerment Measure (GEM). These functions are designed to be user-friendly, requiring minimal coding effort while ensuring accurate and reliable computations.
 
 
 # Overview: <img src="https://github.com/mashrur-ayon/genderstat/blob/main/plots-pictures/genderstat-logo.png" align="right" height="200">
-The `genderstat` R package is a comprehensive tool designed to facilitate quantitative analysis in the realm of gender studies. Developed by S M Mashrur Arafin Ayon, this package stands at version 0.1.3 and serves as an invaluable asset for researchers, students, and professionals in gender studies.
+The `genderstat` R package is a useful tool for quantitative analysis in gender studies. Developed by S M Mashrur Arafin Ayon, this package (version 0.1.5) is intended for researchers, students, and professionals working in gender studies.
+
 
 ## Features:
-- **Functions for Key Gender Metrics:** The package boasts an array of functions tailored to compute various gender inequality metrics. Notably, it can calculate the Gender Pay Gap, Gender Inequality Index (GII), Gender Development Index (GDI), and the Gender Empowerment Measure (GEM).
+**Functions for Key Gender Metrics:** The package includes functions to calculate important gender inequality metrics such as the Gender Pay Gap, Gender Inequality Index (GII), Gender Development Index (GDI), and Gender Empowerment Measure (GEM).
 
-- **Inclusive Datasets:** genderstat is enriched with real datasets extracted from reputable sources, specifically the *UNDP Human Development Reports Data Center* and the *World Bank Gender Data Portal*. These datasets serve as a practical resource for hands-on practice and in-depth learning.
+# Installation
+To install the genderstat package, you have two options: installing from CRAN or directly from GitHub.
 
-- **Dependencies and Utilities:** The package leverages the capabilities of other notable R packages such as `ggplot2`, `dplyr`, and `reshape2`. For those keen on testing and documentation, `testthat`, `knitr`, and `rmarkdown` are suggested.
+To install the stable release from CRAN, use the following command in your R console:
 
-Installation
-To use `genderstat` effectively, you'll also need the `tidyverse` and `ggplot2` packages:
-
-```
-install.packages("tidyverse")
-install.packages("ggplot2")
+```r
 install.packages("genderstat")
-Loading Required Packages
 ```
 
+For the latest development version, you can install the package directly from GitHub using the `devtools` package. First, ensure that `devtools` is installed:
+
+```r
+install.packages("devtools")
 ```
-library(tidyverse)
-library(genderstat)
+
+Then, use the following command to install `genderstat` from GitHub:
+
+```r 
[…]
