### 2024-07-07 mashrur-ayon/genderstat#5 @mashrur-ayon
I have resolved the issue, and in the new version I have removed the need for `ggplot2` and `tidyverse`. A major reason for this is because, after taking into consideration of your suggestion I have removed the plotting functions from this version of `genderstat`. Maybe I will create another package with just the plotting function later on.
