+++ genderstat/.github/workflows/R-CMD-check.yaml
@@ -1,0 +1,24 @@
+name: R-CMD-check
+
+on: [push, pull_request]
+
+jobs:
+  R-CMD-check:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Set up R
+        uses: r-lib/actions/setup-r@v2
+
+      - name: Install dependencies
+        run: |
+          install.packages('remotes')
+          remotes::install_deps(dependencies = TRUE)
+          remotes::install_cran('testthat')
+
+      - name: Check package
+        run: R CMD check --no-manual --as-cran
+
+      - name: Run tests
+        run: Rscript -e 'testthat::test_dir("tests/testthat")'
