### 2024-07-07 mashrur-ayon/genderstat#6 @mashrur-ayon
I have implemented a test using `testthat` as per your suggestion.
