### 2024-07-07 2091dee
update
yaml file update for workflow, and dimension of the figures.
