+++ man/real_data_GDI.Rd
@@ -1,0 +1,27 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/gender_development_index.R
+\docType{data}
+\name{real_data_GDI}
+\alias{real_data_GDI}
+\title{Original Data of GDI (Gender Development Index)}
+\format{
+A data frame with 191 rows and 7 columns:
+\describe{
+  \item{country}{A character vector representing the name of the country.}
+  \item{female_life_expectancy}{A numeric vector representing the life expectancy of females in the respective country, measured in years.}
+  \item{male_life_expectancy}{A numeric vector representing the life expectancy of males in the respective country, measured in years.}
+  \item{female_mean_schooling}{A numeric vector representing the average number of schooling years for females in the respective country.}
+  \item{male_mean_schooling}{A numeric vector representing the average number of schooling years for males in the respective country.}
+  \item{female_gni_per_capita}{A numeric vector representing the Gross National Income per capita for females in the respective country, measured in international dollars.}
+  \item{male_gni_per_capita}{A numeric vector representing the Gross National Income per capita for males in the respective country, measured in international dollars.}
+}
+}
+\source{
+Data obtained from \href{https://www.kaggle.com/dsv/6359326}{Kaggle}.
+}
+\usage{
+data(real_data_GDI)
+}
+\description{
+A dataset containing observed values for the GDI analysis. This data includes metrics for evaluating gender disparities in health, education, and economic resources across different countries.
+}
