### 2024-07-07 #5932 @mashrur-ayon
Dear Reviewers,

@thegargiulian and @joaojcorreia

I hope this message finds you well. I am writing to inform you that I have addressed all the comments and suggestions provided in your feedback for the paper submission on the `genderstat` package. I appreciate your patience and understanding as I took a bit more time to thoroughly implement the necessary changes.

**Summary of Changes:**

1. **Clarified Purpose and Data Sources:**
   - In the paper and the package I have mentioned that the dataset used was taken from secondary source, and linked the dataset with a DOI, the dataset will be updated on a yearly basis by me.

2. **Simplified Language:**
   - Simplified the language across the paper for better readability while maintaining an academic tone.
   - Removed redundant mentions of the package's functionalities to streamline the content.

3. **Detailed Statement of Need:**
   - Expanded the statement of need to include theoretical concepts and the necessity of quantitative tools in gender studies.
   - Explained the gap in existing software solutions and how `genderstat` fills this gap by providing a dedicated tool for gender disparity analysis.

4. **Core Theoretical Concepts:**
   - Provided detailed explanations for the Gender Pay Gap (GPG), Gender Inequality Index (GII), Gender Development Index (GDI), and Gender Empowerment Measure (GEM).
   - Discussed the rationale behind choosing these specific metrics.

5. **Application Section:**
   - Added a comprehensive application section with case studies showcasing the practical use of the package for analyzing gender disparities.
   - Included code examples and visualizations to illustrate the functionality of the package.

6. **Updated Documentation:**
   - Ensured all functions are well-documented with clear examples.
   - Updated the README file on GitHub to reflect the latest changes and included installation instructions for both CRAN and GitHub versions.

7. **Testing and Continuous Integration:**
   - Implemented a unit test using the `testthat` package to ensure the correctness of calculations.

8. **Bibliography and References:**
   - Updated the bibliography to include all references in the correct format, ensuring all sources are properly cited.

I sincerely thank you for your valuable feedback, which has significantly improved the quality and robustness of the paper and the `genderstat` package. Your insights were instrumental in guiding these revisions.

Please find the revised manuscript attached for your review. I look forward to any further feedback you may have and am committed to making any additional improvements as necessary.

Thank you once again for your time and effort.

Best regards,

S M Mashrur Arafin Ayon
