### 2024-07-07 #5932 @mashrur-ayon
Kindly accept this updated pdf
