+++ genderstat/DESCRIPTION
@@ -15,3 +15,5 @@ Suggests:
 Config/testthat/edition: 3
 VignetteBuilder: knitr
 Depends: R (>= 3.5.0)
+License: MIT + file LICENSE
+
