### 2024-07-09 dff430b
adjusted the license file
