+++ genderstat/DESCRIPTION
@@ -15,5 +15,4 @@ Suggests:
 Config/testthat/edition: 3
 VignetteBuilder: knitr
 Depends: R (>= 3.5.0)
-License: MIT + file LICENSE
 
