### 2024-07-11 #5932 @arfon
My apologies to all involved here for triggering this process mid-way through the review (rather than the pre-review issue) but I am concerned about whether this submission meets our[substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion.

I'm going to trigger a scope review and will return in the coming days with a response from the wider editorial team.
