### 2024-07-11 #5932 @arfon
@mashrur-ayon – after consideration by the JOSS editorial team, this submission has been determined as out of scope for JOSS, failing to meet our [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion.

Again, my apologies to all involved here (especially the reviewers @thegargiulian & @joaojcorreia), this submission should have been flagged at pre-review, and as a result not been sent out for review.

@mashrur-ayon – One possible alternative to JOSS is to follow [GitHub's guide](https://guides.github.com/activities/citable-code/) on how to create a permanent archive and DOI for your software. This DOI can then be used by others to cite your work.
