### 2024-07-12 #5932 @mashrur-ayon
> My apologies to all involved here for triggering this process mid-way through the review (rather than the pre-review issue) but I am concerned about whether this submission meets our[substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion.
>
> I'm going to trigger a scope review and will return in the coming days with a response from the wider editorial team.

@arfon To address your concerns adequately, **could you please clarify which specific aspects of the genderstat package do not meet the substantial scholarly effort criterion**?? Understanding the details regarding the commit history, lines of code, number of commits, or any other factors will help me provide the necessary information or make the required adjustments, and resubmit the paper.
