### 2024-07-12 #5932 @arfon
JOSS has a desk rejection policy for submissions totalling under 300 lines of code, but in reality, typically anything under ~1000 LOC is flagged for review.

I will also note that the capabilities of this package, while potentially useful, are also very limited which would categorise this as a “Minor utility” package.
